requirements.txt
ultralytics
opencv-python
rich
hydra-core
Pillow